requirements.txt
# Core Framework
crewai>=0.28.0
crewai-tools>=0.2.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-groq>=0.1.0

# Embeddings & ML
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0

# Vector Database
chromadb>=0.4.0

# APIs & Search
tavily-python>=0.3.0

# Web Framework
streamlit>=1.30.0

# Document Generation
reportlab>=4.0.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging & Monitoring
colorlog>=6.8.0
